requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pypdf>=3.0.0
pdf2image>=1.16.3
pillow>=9.5.0
opencv-python-headless>=4.7.0
pytesseract>=0.3.10
pdfplumber>=0.11.0
chromadb>=0.5.0
openai>=1.40.0
sqlmodel>=0.0.21
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.1
pydantic>=2.8.0
pydantic-settings>=2.4.0
python-multipart>=0.0.9
transformers>=4.35.0
torch>=2.2.0
sentence-transformers>=2.2.2
sentencepiece>=0.1.99
# Optional / advanced OCR (PaddleOCR recommended for handwriting):
paddleocr>=2.6.1
# NOTE: paddlepaddle must be installed separately (CPU or GPU wheel) — see instructions below
email-validator>=2.1.0
python-magic>=0.4.27
